simplify tests, move order_descriptions fixture to conftest
